Refactor to merge PushProjectionThroughLimit and PushProjectionThroughOffset
